comment_analyzer/requirements.txt
Django==4.1.4
Jinja2==3.1.2
emoji==2.2.0
matplotlib==3.6.2
matplotlib-venn==0.11.7
mpld3==0.5.9
pandas==1.5.2
plotly==5.11.0
youtube-comment-downloader==0.1.67
youtube-search-python==1.6.6
stop-words==2018.7.23
django-simple-captcha==0.5.17
python-slugify==8.0.0
djoser==2.1.0
django-debug-toolbar==3.8.1